requirements.txt
nats-py>=2.0.0